#BLEM-63 Add Feedback Section in user tab (#65)
* Make Model Feedback --all

* Fix DB migrations

* Update Enums

* Refactor Form Components

* Delete unused components

* Update components

* Savepoint 01

* Add Feedback list

* Update Feedback list

* Ui Update

* Ui Update, Add Rating funcionality

* Update Feedback validation

* `seen` feature

+++ database/migrations/2023_04_07_170447_create_acquaintances_table.php
@@ -11,25 +11,30 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('acquaintances', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('transmitter_user_id')
-                ->constrained('users')
-                ->comment('the user that initialised the friend request')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->foreignId('receiver_user_id')
-                ->constrained('users')
-                ->comment('the user that received the friend request')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->enum('status', ['pending', 'accepted','denied','blocked'])
-                ->default('pending')
-                ->comment('pending/accepted/denied/blocked');
-            $table->timestamps();
-            // composite unique key
-            $table->unique(['transmitter_user_id', 'receiver_user_id']);
-        });
+        if(!Schema::hasTable('acquaintances'))
+            Schema::create('acquaintances', function (Blueprint $table)
+            {
+                $table->id();
+
+                $table->foreignId('transmitter_user_id')
+                    ->constrained('users')
+                    ->comment('the user that initialised the friend request')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->foreignId('receiver_user_id')
+                    ->constrained('users')
+                    ->comment('the user that received the friend request')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->enum('status', ['pending', 'accepted','denied','blocked'])
+                    ->default('pending')
+                    ->comment('pending/accepted/denied/blocked');
+
+                $table->timestamps();
+
+                // composite unique key
+                $table->unique(['transmitter_user_id', 'receiver_user_id']);
+            });
     }
 
     /**
